fix: remove InputField, and move Input into core components
