--- web/app/components/InputField.vue
@@ -1,14 +1,0 @@
-<template>
-  <input v-model="model" :placeholder="placeholder" />
-</template>
-
-<script setup>
-const model = defineModel();
-
-const props = defineProps({
-  placeholder: {
-    type: String,
-    default: null,
-  },
-});
-</script>
