+++ web/app/pages/admin/[location]/products/components/FilterBar.vue
@@ -23,10 +23,11 @@
       <option value="description">{{ t("no_description") }}</option>
     </Select>
     <!-- Text filter -->
-    <InputField
-      v-model="query"
+    <Input
+      id="admin-product-input"
       :placeholder="`${t('search')}...`"
-      class="text-search lb-input"
+      v-model="query"
+      class="text-search"
       @input="$emit('input')"
     />
     <!-- Clear filter button -->
@@ -50,8 +51,8 @@
 <script lang="ts" setup>
 import Button from "@/components/core/Button.vue";
 import HorizontalScroll from "@/components/core/HorizontalScroll.vue";
+import Input from "@/components/core/Input.vue";
 import Select from "@/components/core/Select.vue";
-import InputField from "@/components/InputField.vue";
 import { Xmark } from "@iconoir/vue";
 
 const { t } = useI18n({
@@ -70,7 +71,7 @@ defineEmits<{ input: [] }>();
   gap: var(--fluid-spacing-4);
 }
 .text-search {
-  width: auto;
+  width: auto !important;
 }
 .no-photos {
   width: auto;
