+++ web/app/pages/admin/[location]/products/components/ProductDrawer.vue
@@ -120,10 +120,10 @@ import Button from "@/components/core/Button.vue";
 import Dialog from "@/components/core/Dialog.vue";
 import Drawer from "@/components/core/Drawer.vue";
 import Heading from "@/components/core/Heading.vue";
+import Input from "@/components/core/Input.vue";
 import RichTextarea from "@/components/core/RichTextarea.vue";
 import Switch from "@/components/core/Switch.vue";
 import ImageInput from "@/components/ImageInput.vue";
-import Input from "@/components/Input.vue";
 import { Eye, Trash, Xmark } from "@iconoir/vue";
 import type { RecordModel } from "pocketbase";
