+++ web/app/pages/l/[location]/p/[product].vue
@@ -173,10 +173,10 @@ import Container from "@/components/core/Container.vue";
 import DateInput from "@/components/core/DateInput.vue";
 import Dialog from "@/components/core/Dialog.vue";
 import Heading from "@/components/core/Heading.vue";
+import Input from "@/components/core/Input.vue";
 import KeyValue from "@/components/core/KeyValue.vue";
 import Textarea from "@/components/core/Textarea.vue";
 import Tooltip from "@/components/core/Tooltip.vue";
-import Input from "@/components/Input.vue";
 import PageAlert from "@/components/page-alert/PageAlert.vue";
 import ProductImage from "@/components/ProductImage.vue";
 import ReservationsBox from "@/components/ReservationsBox.vue";
