+++ web/app/pages/login.vue
@@ -63,7 +63,7 @@ import Button from "@/components/core/Button.vue";
 import Card from "@/components/core/Card.vue";
 import Container from "@/components/core/Container.vue";
 import Heading from "@/components/core/Heading.vue";
-import Input from "@/components/Input.vue";
+import Input from "@/components/core/Input.vue";
 import {
   AFTER_LOGIN,
   AFTER_LOGIN_RESERVATION_INTENT,
