+++ web/app/pages/reset-password.vue
@@ -33,7 +33,7 @@ import Button from "@/components/core/Button.vue";
 import Card from "@/components/core/Card.vue";
 import Container from "@/components/core/Container.vue";
 import Heading from "@/components/core/Heading.vue";
-import Input from "@/components/Input.vue";
+import Input from "@/components/core/Input.vue";
 
 const { t } = useI18n({
   useScope: "local",
